Move dirty and readonly flags to the context.
