+++ packages/docregistry/src/mimedocument.ts
@@ -191,6 +191,7 @@ export class MimeContent extends Widget {
     } else {
       if (!JSONExt.deepEqual(data, this._context.model.toJSON())) {
         this._context.model.fromJSON(data);
+        this._context.dirty = false;
       }
     }
   };
